Only delete CMake cache file if it exists.

+++ scripts/build.sh
@@ -8,7 +8,7 @@ if [ ! -d build ]; then mkdir build; fi
 
 cd build
 
-rm CMakeCache.txt
+if [ -f CMakeCache.txt ]; then rm CMakeCache.txt; fi
 
 if grep -q 'Raspbian' /etc/os-release; then
   cmake ..
